Adding compute and updating vertex indices to use unsigned 16 bit ints to be compatible with latest Flutter dev channel.

+++ flare_flutter/lib/flare.dart
@@ -676,8 +676,8 @@ class FlutterActor extends Actor {
     if (_rawAtlasData == null) {
       return false;
     }
-	List<Uint8List> data = _rawAtlasData;
-	_rawAtlasData = null;
+    List<Uint8List> data = _rawAtlasData;
+    _rawAtlasData = null;
     List<ui.Codec> codecs =
         await Future.wait(data.map(ui.instantiateImageCodec));
     List<ui.FrameInfo> frames =
@@ -911,7 +911,7 @@ class FlutterActorImage extends ActorImage with FlutterActorDrawable {
   Float32List _uvBuffer;
   ui.Paint _paint;
   ui.Vertices _canvasVertices;
-  Int32List _indices;
+  Uint16List _indices;
 
   void onPaintUpdated(ui.Paint paint) {}
   final Float64List _identityMatrix = Float64List.fromList(<double>[
@@ -972,7 +972,7 @@ class FlutterActorImage extends ActorImage with FlutterActorDrawable {
     _vertexBuffer = makeVertexPositionBuffer();
     _uvBuffer = makeVertexUVBuffer();
     _indices =
-        Int32List.fromList(triangles); // nima runtime loads 16 bit indices
+        Uint16List.fromList(triangles); // nima runtime loads 16 bit indices
     updateVertexUVBuffer(_uvBuffer);
     int count = vertexCount;
     int idx = 0;
@@ -1048,7 +1048,8 @@ class FlutterActorImage extends ActorImage with FlutterActorDrawable {
       }
     }
 
-    _paint.color = _paint.color.withOpacity(renderOpacity.clamp(0.0, 1.0).toDouble());
+    _paint.color =
+        _paint.color.withOpacity(renderOpacity.clamp(0.0, 1.0).toDouble());
 
     if (imageTransform != null) {
       canvas.transform(imageTransform.mat4);
